Move some non-reusable code
