+++ WordPress/Classes/Extensions/Atomic.swift
