--- WordPress/Classes/Extensions/UICollectionViewCell+Tint.swift
@@ -1,8 +1,0 @@
-import UIKit
-
-extension UICollectionViewCell {
-    // Allows cell tint color to be set via UIAppearance
-    @objc func setCellTintColor(_ color: UIColor) {
-        tintColor = color
-    }
-}
